rename: type -> types
